Extract NodeActionButton
